+++ src/guardrails/index.ts
@@ -1,0 +1,10 @@
+/**
+ * SDK Guardrails Module
+ * Layer 7: Verified Data Structure Enforcement
+ * 
+ * @packageDocumentation
+ * @module @roALAB1/manus-validation-kit
+ */
+
+export { SDKGuardrailsEngine, default } from './SDKGuardrailsEngine';
+export * from '../types/sdk-guardrails';